Update Xcode projects to work with mame0106u13. Although it doesn't quite work yet...

+++ macintosh/Source/osinline.h
@@ -49,7 +49,7 @@ INLINE int _vec_mult(int x, int y)
 #endif
 
 /* profiler callback */
-extern unsigned int (*gGetCyclesFunction)(void);
+extern long long (*gGetCyclesFunction)(void);
 #define osd_cycles (*gGetCyclesFunction)
